Add basic benchmark for tracing overhead
Right now the overhead seems to be around 130 ns per call on Mac,
which is consistent with what we know about the overhead of
System.nanoTime(). The overhead will probably be higher on Windows.

Eventually we should also consider measuring tracing overhead in a
multi-threaded environment, because System.nanoTime() can be slower
when multiple threads call it simultaneously.