Handle NoClassDefFoundError during completion
It turns out that methods like Class.getCanonicalName() and
Class.isAnonymousClass() may throw NoClassDefFoundError in some cases.
In fact, this caused auto-completion to break completely when using
ide-perf in Android Studio (NoClassDefFoundError is thrown for
com/android/tools/idea/common/error/IssueSource$DefaultImpls).

This commit adds some exception handling to work around this.
It also updates the filter for which classes to include in completion.