Disentangle the three tracers
The three tracers do not need to share base classes because they
do not have very much code in common. The shared base classes
were making it difficult to understand and evolve the code for
each individual tracer.

Also, command line history is removed for now. This should be
reimplemented using ComboBox and PropertiesComponent.