Avoid using COMPUTE_FRAMES
COMPUTE_FRAMES forces ClassWriter to use reflection, which fails quite
commonly when running IntelliJ because it does not have the right
classloader for the classes it is trying to load.

For example, ide-perf could not trace the AccessibilityDetector class
because it would result in the following error:
    Failed to instrument class
        com/android/tools/lint/checks/AccessibilityDetector
    Type com/android/tools/lint/detector/api/LintFix not present

Plus, leaving the original stack map frames intact improves
performance slightly.